requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-jwt